Move utils, add puck to individual permit pages
