+++ src/app/development/permits/PermitsTable.js
@@ -5,16 +5,14 @@ import Measure from 'react-measure';
 import moment from 'moment';
 import expandingRows from '../../../shared/react_table_hoc/ExpandingRows';
 import createFilterRenderer from '../../../shared/FilterRenderer';
-
+import { defaultTableHeaders } from '../utils';
 
 const ExpandableAccessibleReactTable = expandingRows(AccessibleReactTable);
+
 
 class PermitsTable extends React.Component {
   constructor(props) {
     super(props);
-    // TODO: USE WINDOW.LOCATION INSTEAD OF REACT ROUTER LOCATION?
-    // https://medium.com/@ivantsov/using-react-router-and-history-38c021270829
-
     const filtered = [];
     if (window.location.search) {
       const queryParams = window.location.search.slice(1).split('&');
@@ -74,7 +72,7 @@ class PermitsTable extends React.Component {
                     id: headerObj.field,
                     accessor: (d) => {
                       return headerObj.formatFunc ?
-                        headerObj.formatFunc(d[headerObj.field]) :
+                        headerObj.formatFunc(d) :
                         d[headerObj.field];
                     },
                     Filter: createFilterRenderer(`Search ${headerObj.display}`),
@@ -130,37 +128,7 @@ PermitsTable.propTypes = {
 
 PermitsTable.defaultProps = {
   data: [],
-  tableHeaders: [
-    {
-      field: 'applied_date',
-      display: 'Date Applied',
-      formatFunc: d => moment.utc(new Date(d)).format('MMM DD, YYYY'),
-    },
-    {
-      field: 'address',
-      display: 'Address',
-    },
-    {
-      field: 'permit_subtype',
-      display: 'Type',
-      show: colWidth => colWidth > 70,
-    },
-    {
-      field: 'status_current',
-      display: 'Status',
-      show: colWidth => colWidth > 90,
-    },
-    {
-      field: 'applicant_name',
-      display: 'Applicant',
-      show: colWidth => colWidth > 90,
-    },
-    {
-      field: 'permit_number',
-      display: 'Record Link',
-      formatFunc: d => <a href={`/permits/${d}`}>{d}</a>,
-    },
-  ],
+  tableHeaders: defaultTableHeaders,
 };
 
 export default PermitsTable;
