+++ src/app/development/trc/MajorDevelopmentDashboard.js
@@ -4,24 +4,6 @@ import SectionNav from './SectionNav';
 import AnnotatedDagre from './AnnotatedDagre';
 import TRCDataTable from './TRCDataTable';
 import Accordion from '../../../shared/visualization/Accordion';
-import { trcProjectTypes } from './utils';
-
-const projectTypes = trcProjectTypes;
-
-// Assign colors to project types
-const orderedColors = [
-  '#FF3A3A',
-  '#749B5F',
-  '#2d93ad',
-  '#004EA3',
-  '#9B6681',
-  '#9E4F55',
-  '#073d49',
-];
-Object.keys(projectTypes).forEach((type, i) => {
-  projectTypes[type].color = orderedColors[i];
-});
-
 
 class MajorDevelopmentDashboard extends React.Component {
   constructor() {
@@ -45,9 +27,7 @@ class MajorDevelopmentDashboard extends React.Component {
             <p>After the developer submits an application, it goes through a decision-making process that includes city staff, city council, developers, and residents.  Who is involved at what step depends on the type of project.</p>
             <p>The Unified Development Ordinance defines six types of large scale development in Asheville.</p>
             <p>Depending on the type of project, there are three to eight steps.</p>
-            <AnnotatedDagre
-              projectTypes={projectTypes}
-            />
+            <AnnotatedDagre />
           </div>
         ),
       },
