+++ src/styles/components/MajorDevelopmentDashboard.scss
@@ -34,11 +34,10 @@ ul.sectionNav li {
   display: flex;
   border-bottom: 1px solid #f2f2f2;
   width: 100%;
-  font-size: 0.75em;
 }
 
 .sectionNav a {
-  padding: 0.75rem 0.25rem 0.75rem 0.75rem;
+  padding: 0.75rem 0.25rem 0.75rem 1rem;
   letter-spacing: 0.035rem;
   width: 100%;
 }
